Place build target in build root

+++ build.sh
@@ -33,6 +33,7 @@ ULTRASCHALL_SCRIPTS_DIRECTORY="./scripts"
 ULTRASCHALL_DOCS_DIRECTORY="./docs"
 
 ULTRASCHALL_BUILD_OUTPUT="html"
+ULTRASCHALL_BUILD_TARGET="ultraschall-manual.$ULTRASCHALL_BUILD_OUTPUT"
 
 ULTRASCHALL_BUILD_BOOTSTRAP=0
 ULTRASCHALL_BUILD_CLEAN=0
@@ -75,21 +76,37 @@ echo "*                                                                    *"
 echo "**********************************************************************"
 echo ""
 
-if [ $ULTRASCHALL_BUILD_CLEAN -ne 0 ]; then
-  echo "Cleaning Ultraschall manual files..."
-  rm -rf $ULTRASCHALL_BUILD_DIRECTORY
-  echo "Done."
+if [ ! -x "$(command -v pandoc)" ]; then
+  echo "ERROR: The build script requires the pandoc command."
+  ULTRASCHALL_BUILD_FAILED=1
 fi
 
-if [ ! -d $ULTRASCHALL_BUILD_DIRECTORY ]; then
-  mkdir -p $ULTRASCHALL_BUILD_DIRECTORY
+if [ $ULTRASCHALL_BUILD_FAILED -eq 0 ]; then
+  if [ $ULTRASCHALL_BUILD_CLEAN -ne 0 ]; then
+    echo "Cleaning Ultraschall manual files..."
+    rm -rf $ULTRASCHALL_BUILD_DIRECTORY
+    echo "Done."
+  fi
+fi
+
+if [ $ULTRASCHALL_BUILD_FAILED -eq 0 ]; then
+  if [ ! -d $ULTRASCHALL_BUILD_DIRECTORY ]; then
+    mkdir -p $ULTRASCHALL_BUILD_DIRECTORY
+  fi
 fi
 
 if [ $ULTRASCHALL_BUILD_FAILED -eq 0 ]; then
   echo "Building Ultraschall manual files..."
-  pandoc --from=markdown --to=$ULTRASCHALL_BUILD_OUTPUT --standalone --self-contained --quiet --css=$ULTRASCHALL_SCRIPTS_DIRECTORY/ultraschall.css --output=$ULTRASCHALL_BUILD_DIRECTORY/ultraschall-manual.html $ULTRASCHALL_DOCS_DIRECTORY/outline.md
+  pandoc --from=markdown --to=$ULTRASCHALL_BUILD_OUTPUT --standalone --self-contained --quiet \
+    --css=$ULTRASCHALL_SCRIPTS_DIRECTORY/ultraschall.css \
+    --output=$ULTRASCHALL_BUILD_TARGET \
+    $ULTRASCHALL_DOCS_DIRECTORY/outline.md
   if [ $? -ne 0 ]; then
     ULTRASCHALL_BUILD_FAILED=1
   fi
   echo "Done."
+fi
+
+if [ $ULTRASCHALL_BUILD_FAILED -ne 0 ]; then
+  echo "Failed to build $ULTRASCHALL_BUILD_TARGET."
 fi
